requirements.txt
open_clip_torch
torch
Pillow
matplotlib